requirements.txt
langchain
python-dotenv
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain_openai
chromadb
sentence_transformers
duckdb
faiss-cpu
langchain_huggingface
langchain_chroma
pandas
openai
langchain-groq
duckduckgo-search
langchain_community
numexpr
pytube
unstructured
youtube_transcript_api
mysql-connector-python
SQLALchemy
validators==0.28.1
streamlit
pypdf